Add files via upload
Predicting Churn User or Customer attrition Project is based on Identifying which Customer is going to Quit using the services in the near future. With the help of Advancing technology like python, we perform Exploratory Data analysis on the data and understand it with Data Visualization and with Machine Learning implementation of models is concluded and also, we able to find the key points affecting the user to churn.